fix(windows): repair media controls build
Bump version to v0.1.12 and gate platform-specific imports to avoid Windows CI warnings.

+++ src/integrations/media_controls/native.rs
@@ -243,7 +243,7 @@ impl Drop for HiddenWindow {
         use windows::Win32::{Foundation::HWND, UI::WindowsAndMessaging::DestroyWindow};
 
         unsafe {
-            let _ = DestroyWindow(HWND(self.hwnd as isize));
+            let _ = DestroyWindow(HWND(self.hwnd));
         }
     }
 }
